requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6

# AI/ML dependencies
torch==2.1.0
transformers==4.35.0
accelerate==0.24.1

# LangGraph and LangChain
langgraph==0.0.62
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# Frontend dependencies
streamlit==1.28.1
streamlit-option-menu==0.3.6

# Data processing
pandas==2.1.3
python-dateutil==2.8.2
nltk==3.8.1

# Web scraping and article extraction
newspaper3k==0.2.8
trafilatura==8.4.0
requests-html==0.10.0
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF generation
reportlab==4.0.7
weasyprint==60.2

# Caching
diskcache==5.6.3

# Environment variables
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Additional dependencies for LangGraph
typing-extensions==4.8.0
pydantic==2.5.0
